sec try

+++ hash_tables/5-hash_table_print.c
@@ -9,7 +9,7 @@
  */
 void hash_table_print(const hash_table_t *ht)
 {
-int j = 0;
+unsigned long int j = 0;
 int i = 1;
 hash_node_t *node;
 if (ht == NULL)
